Exposing contract facade
